chore: Create chat domain folder and reorganize related files
- Created a chat folder to organize chat-related functionality.
- Moved relevant files into the new folder
