+++ client/src/components/chat/ChatBubbleUI.tsx
